requirements.txt
blessings 
brotlipy==0.7.0
certifi 
cffi 
charset-normalizer 
cryptography 
cycler==0.11.0
fonttools==4.34.4
gpustat 
idna 
imageio==2.19.3
kiwisolver==1.4.3
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
networkx==2.8.4
numpy 
nvidia-ml-py3 
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
psutil 
pycparser 
pyOpenSSL 
pyparsing==3.0.9
PySocks 
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests 
scikit-image==0.19.3
scipy==1.8.1
six 
thop==0.1.0.post2207010342
tifffile==2022.5.4
timm==0.6.5
tk==0.1.0
torch==1.12.1
torchstat==0.0.7
torchvision==0.13.1
typing_extensions 
urllib3 
